requirements.txt
tensorflow==2.19.0
numpy>=1.21.0
Pillow>=8.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0